+++ app/src/main/java/org/akvo/flow/util/ConstantUtil.java
@@ -114,9 +114,9 @@ public class ConstantUtil {
     public static final String FORM_ID_EXTRA = "SID";
     public static final String RESPONDENT_ID_EXTRA = "survey_respondent_id";
     public static final String READ_ONLY_EXTRA = "readonly";
-    public static final String SURVEY_GROUP_EXTRA = "survey_group";
+    public static final String SURVEY_EXTRA = "survey_group";
     public static final String DATA_POINT_ID_EXTRA = "datapoint_id";
-    public static final String SURVEY_GROUP_ID_EXTRA = "survey_group_id";
+    public static final String SURVEY_ID_EXTRA = "survey_group_id";
     public static final String IMAGE_URL_EXTRA = "image_url";
     public static final String FORM_TITLE_EXTRA = "title";
     public static final String FORM_SUBTITLE_EXTRA = "subtitle";
@@ -214,6 +214,7 @@ public class ConstantUtil {
     public static final int STORAGE_PERMISSION_CODE = 2;
     public static final int STORAGE_AND_PHONE_STATE_PERMISSION_CODE = 3;
 
+    public static final String LANGUAGES_EXTRA = "languages";
 
     /**
      * prevent instantiation